EventHub/tests/requirements.txt
# For send_event.py (E2E test)
azure-eventhub>=5.0.0
